Contact taskd server on new account creation
This validates the connection parameters are correct.

Add the taskwarrior-java-client library as a dependency to communicate
with a taskd server.

There are currently some limitations with the underlying library:
- server must be passed as IP address rather than FQDN
- private key certificate must be DER encoded

To convert a PEM key to DER, use the following command:
openssl pkcs8 -topk8 -nocrypt -in key.pem -inform PEM -out key.der -outform DER